[FIX] UI fix to Create Pupil and Edit Pupil. "Fill all fields" error fixed.

+++ app/src/main/java/com/bridge/androidtechnicaltest/db/PupilDao.java
@@ -40,6 +40,6 @@ public interface PupilDao {
     void clearAll();
 
     @Query("SELECT * FROM Pupils WHERE pupil_id = :id")
-    Single<Pupil> getPupilById(int id);
+    Single<Pupil> getPupilById(Long id);
 
 }
